Add flake8 linting and black formatting to CI
Currently we have flake8 configured as pre-push hook but not configured in
tox which gets ran in Prow. [1]

This change adds flake8 linting to the tox config, updates the tox Makefile
target to run the flake8 job and fixes errors reported by flake8.

This change also enables the pep8 `max-line-length` check and bumps it
to 88 as recommended by black[2]

This change also enables the black formatter in CI and pre-commit.

[1] https://github.com/openshift/release/blob/f0d768950a3510ca1a9ee7d79b476bca1c9d8e8e/ci-operator/config/openstack-k8s-operators/ci-framework/openstack-k8s-operators-ci-framework-main.yaml#L38-L41
[2] https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#line-length