Do not set cifmw_cephadm_first_mon_ip from Ceph spec
Fixes bug from e23977349353fad521dd4ded0f149d3cab18c6fd

The "Set IP address of first monitor" task broke when
the Ceph spec was changed to always use the ssh_network
because that network is not necessarily the same as the
storage_network_range. Update this task to get the first
monitor IP not from the spec but from the hostvars of the
host where the play is run after filtering for only the
storage network.

Signed-off-by: John Fulton <fulton@redhat.com>