New role: ci_network
This role will apply network configuration on the hosts.

While it has a default network layout, it is expected to
get data from a previous step, such as the Zuul job bootstrap,
in a specific format matching the default layout.

In the case of the job reproducer, the data will be generated by mixing
data from the actual CI job, and the ones matching the reproducer
infra.