Fix CoreDNS loading 127.0.0.1 as nameserver
CoreDNS is configured by OCP to use the crc /etc/resolv.conf as the main
resolvers source, but since it points to 127.0.0.1 CoreDNS tries to use
itself as a resolver instead of the host dnsmasq. To avoid that we
simply points the resolv.conf entry to the isolated net IP of crc
instead of localhost.
At the sametime, some other hardcoded values are changed to use the
values from the networking file.