Fix network authold logic
Some changed to fix how the network autohold mechanism works. The API
URL was wrong and the field used to compare the project name should be
the canonical one. Fixed a typo in the filename too.
As successful runs don't tigger the instances hold we now skip holding
the network for successful runs.