Address comments from #586
As discussed in the mentioned PR, a follow-up had to be created in order
to address some comments.

This PR corrects a small typo in a comment/doc, and introduce a new task
file for the network layout validation.

We also take the opportunity to remove the useless "vars" directory.