Set DATAPLANE_GROWVOLS_ARGS for baremetal job
Now that DATAPLANE_GROWVOLS_ARGS is a variable the current value should
be set in the job, since the value is specific to the
BMAAS_INSTANCE_DISK_SIZE in this job.

The value in install_yamls[1] can be set to an empty string after this
is merged, so that developers can install with the default.

[1] https://github.com/openstack-k8s-operators/install_yamls/blob/main/devsetup/Makefile#L34