Enable new parameter support for compute hook
Since we intend to support IPv6 sooner or later, it's better to be
consistent with the parameter names. Since the ci_network role
introduces the "ip6" and "ip4" as well as "dns6" and "dns4", let's make
the hook aware of them for the reproducer case.

This will also make the switch to ci_network role easier in the near
future.

Regarding the "!crc,!crc": those exclusions ensure we don't try to
deploy repositories on the CRC. Both names are mandatory:
- "crc" is the name in the CI
- "crcs" is a group name, present in the reproducer.

While it will raise a warning stating that ansible can't match one of
those entries, it won't fail.