Modify libvirt_manager role for proper reproducer capabilities
This patch extend the libvirt_manager role in order to create the full
reproducer job layout.
In such a case, there are some extended needs compared to the "deploy a
layout" easy use-case: we need to inject some keys, create network
layout files/configuration and so on.

This patch, in conjuction with some others coming later, will allow
users to reproduce a 1:1 CI job, with the layout and networking.