Reproducer: add content-provider support
This patch introduce content-provider job-like support.

The goal here is to enable job reproduction in cases such as the
podified-multinode-edpm-deployment-crc.
This specific job requires a content-provider job in order to build and
provide the operator/container images.

With this patch, such a need is detected by the role, and a step to get
the content-provider running on the ansible controller node is then
launched.

This patch also corrects some wrong assumption with the way zuul
clones/manages the dependency repositories, and is closer to its
behavior than before.

Finaly, we also introduce a new tag, "boostrap_layout" in order to
re-run only specific steps.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/637