No cert validation in os_net_setup
For now disable cert validation on os_net_setup role. In a follow
up we fetch the CA from a CA bundle secret and can enable it. Disable
it is a pre-req to allow integration of the TLS functionality for the
public endpoints via the openstack-operator. The openstack-operator
will create the required CA.