Split edpm_baremetal into 2 tasks
The documented process is to call devsetup/Makefile
edpm_baremetal_compute and then Makefile edpm_deploy_baremetal. This
change does that, and also aligns the task names with the compute
tasks in 06-deploy-edpm.yml.

Once this has landed, target edpm_baremetal can be deleted so there is
no coupling between the 2 Makefiles.