Disable gpgcheck during rhos-release rpm installation
On RHEL system, if we install rhos-release from rpm. It gives
```
Failed to validate GPG signature for rhos-release-1.6.15-1.noarch:
Package rhos-release-latest.noarchj1qd2t_h.rpm is not signed
```
These rpms are not shipped to customers and if the machine is
subscribed subscription manager, we hit the above issue. The only
way to overcome the issue is to disable the gpgcheck during
rhos-release rpm installation.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>