Replace ci_make invocations with ci_script to reduce maintenance
The ci_make calls can be accommodated with ci_script. This replaces
all invocations of ci_make with ci_script and updates tests accordingly.
The params: argument of ci_make has been replaced with extra_args: that
has been added to the ci_script module.

The removal of the ci_make module and associated files will come
in followup reviews.

https://issues.redhat.com/browse/OSPRH-1630 (OSP-29266)