Add high level tags to deploy-edpm.yml playbook
If the deploy-edpm.yml playbook fails it saves time to be
able to re-run it and pass tags so certain steps can be
skipped.

Tags are named their plays. The control-plane tag is the
same as the edpm tag for now since the 06-deploy-edpm.yml
playbook also deploys the control-plane. Further tagging
can be done within the edpm_prepare role to differentiate
the control-plane and data-plane.

Tags were not added to the 01-bootstrap.yml playbook call
because the bootstrap tag already exists and the playbook
has some tasks with an always tag.

Signed-off-by: John Fulton <fulton@redhat.com>