[devscripts] Add file related vars for pull secrets and ci-token
Currently we are passing pull secrets and ci-token content as a string
in the ansible var. In order to consume it in CI, we need to slurp
the content of pull secret and ci-token in these vars.

If somehow the these vars are passed via var files through command line,
it might expose the secrets on failures.

This pr adds two new vars to pass the pull secrets and ci-token as file
also.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>