Correct check for ca_bundle_file in edpm_prepare
The task "Inject OpenStackControlplane CA bundle" checks whether the
ca_bundle contents was written to file, but in its current version will
run even if the previous task was skipped, and it was not written to
fail. Logically, the invocation of the install_ca role then fails.

This change modifies the condition of the task so it only runs if the
previous task was changed, so install_ca will not be called if it was
skipped.