Ensure we clean ssh key from the authorized_keys
Until now, in case we were iterating deploy and cleanup over and over,
with the removal of the `cifmw_devscripts_basedir`, we ended up with an
ever-growing authorized_keys.

This patch ensures, during the cleanup tasks, that we remove the
generated key. It won't have impact on the deployed nodes, since, at
this point, there won't be any node anymore.