Improve reproducer role
With this patch, we extend the reproducer role in such a way it will
properly integrate and support dev-scripts provided VMs, and properly
manage the ones we may want to inject using libvirt_manager.

With this change, we ensure that:
- devscripts is running first, so that it can create its networks

- libvirt_manager comes next, and is able to consume any existing
  networks provided by devscripts

- the reproducer is able to generate and provide mandatory files related
  to the inventory and network, even if we're not running a CI Job
  Reproducer

- we introduce a "light" EPEL management, in the wait of some formal
  inclusion in either ci_setup or repo_setup

- we correct a tiny issue in the `push_code.yml` that prevented correct
  matching of the "kind" of repository (remote or local), since `match()`
  is greedy.

This patch is a major refactoring of the reproducer role, allowing to
get extended features.