[set_openstack_containers] filter os containers based on prefix
Openstack operators got the support for disconnected environment. It
changes the openstack services container url to use sha instead of tag.

It breaks the set_openstack_containers role as we were filtering the
containers based on tags.

With this pr, it removes the filter var and replace it with
image prefix and also allow user to use the image prefix in the
image name.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>