Fix kube:admin for OAuth logins
In OCP kube:admin user, the one in the cluster, is mapped to
the OAuth enabled user kubeadmin. whoami, the way we use to
return the user always returns kube:admin for that user, but
that username cannot be used to login using OAuth. This
behaviour is hardcoded inside OCP itself.