Ignore operator related images
Currently set_openstack_containers role filter the openstack services
images based on openstack or downstream_prefix name.

But openstack-baremetal-operator-agent is not a openstack service image.
It needs to be ignored by set_openstack_containers role.

In order to do that, we are adding another conditional check to skip
operator image.

It will fix the job.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>