Allow overlay support in devscripts role
This patch allows to instruct devscripts to stop the cluster right after
it has been created, and so some small tasks such as dumping the domains
XML and so on.

These tasks will allow to describe the OCP cluster via the
libvirt_manager configuration data structure. We'll end up with a "slow"
first deploy, since we're deploying the full OCP, but future iteration
will be really faster since we'll just create an overlay instead of
re-deploying.

Further patches are needed to get full support for that feature.