Avoid warning the user about oc not found
The approach we took for choose if oc needs to be installed or not
is to call it, and if it fails, rescue the task and install it.
That yields an error that a user is usually interpreting as a real
one. We should not pass that error to the user, and ignore it.

At the same time, it seems that under certains circustamces, the fact
catching mechanism is capturing the variable that holds where oc
should be installed, what's a problem if the previous run was done
with a different user, like root, as the extraction of oc will fail
because of lack of permission.