Install tcib from RPM
tcib is used in periodic line to build openstack services
containers.

Periodic line should be instaling the tcib package itself from
podified-ci-testing via rpm. But it is installed via github source.
The tcib content in podified-ci-testing is different from source
(New commit might merged and dlrn have not picked it up.)

So, it diverges the testing and causes unwanted issues.

This pr adds the support for installing the tcib content from
rpm. It will help to install the correct content.

It will be the default behaviour. Installing from source will be
optional.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>