Drop current tags from install_ca role
podified-multinode-edpm-e2e-nobuild-tagged-crc use:
--skip-tags bootstrap,packages
which result in the install_ca role not be triggered when the
created ca bundle got extracted from the deployed env. As a
result the cert validation of the keystone public endpoint
fails.

We should be able to add CAs at any time. Therefore this drops
the current tags from the role.

Jira: https://issues.redhat.com/browse/OSPCIX-104