Create Nova SSH migration secret in hci_prepare role
Add task to phase2 tasks list in hci_prepare role which
prepares the Nova migration SSH key pair.

The podified-multinode-hci-deployment-crc job is failing
ever since the Nova OpenStackDataPlaneSercice started
requiring the Nova migration SSH key secret. The error
was not seen in non-HCI jobs because install_yamls was
updated to automatically handle it [1]. Because the HCI
job is not as tightly coupled with install_yamls, we need
to create this secret as part of the hci_prepare role.

[1] https://github.com/openstack-k8s-operators/install_yamls/pull/586/files

Signed-off-by: John Fulton <fulton@redhat.com>