Allow auth file to be null
Currently the cifmw_build_containers_authfile_path variable
has a default string value and because of that it is impossible
to unset it in Ansible. However, when deploying local registry
with registry_deploy role, the authentication is not necessary
while pushing the built containers images. Instead of supplying
a dummy empty file as workaround, it would be nice to just set
the authfile_path variable to Null/None.

Additionally, a minor typo in task name is corrected.