Fix libvirt_manager and reproducer roles
With all of the changes happening in order to rework the layout, add
support for devscripts and related tasks, it appears the reproducer
broke. More specifically, the network part was out, due to hard-coded
values that leaked following some hackathon (really, hack-a-ton) session
that happened lately.

This patch puts things square again and allows to deploy the needed
bits.

In order to ensure things are working properly, we also do some cleaning
in molecule:
- removing existing tests, since they are mostly the same and are
  consuming install_yamls, with dry_run
- adding a new test that will actually deploy a bunch of VM, consuming
  most of the features available in the role (layout, multi-nic,
  fixed_ips)