Use kustomize add operation for setting insecure registry
https://github.com/openstack-k8s-operators/ci-framework/pull/785 adds
the support for setting insecure registry on EDPM nodes.

It works for EDPM. In BM job, we used replace operation which does not
work as the key `edpm_container_registry_insecure_registries` does not
exists in CR.

We need to use `add` as used in EDPM in order to make it working.

Note: It also bumps the timeout for tcib jobs as new containers got
added and we need more then an hour to build all containers.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>