Get actual test for reproducer role
This patch activate a test for one of the reproducer features:
layout reproducer.

Since it's the base feature of the reproducer role, it's a good idea to
ensure it's working properly.

It also includes some more tweakings in the libvirt_manager role.

We therefore update the nodeset to provide CRC.