Expose a userfriendly container env
ALlowing people to consume the framework via a container seems to be
mandatory.

This change allows to get rid of `buildah` call (not available on MacOS
for instance), while allowing some more friendly usage of the
ci-framework.

The embedded documentation will be added to the readthedocs one once we
finish the deep refacroring.