Switch rhol_crc to the new manage_secrets
This switch should ensure consistency with the whole pull-secret
management.

It will also ensure we're actually passing down the needed bits to the
deploy and throw an early failure if needed.

We also take the opportunity to remove the useless `run:` parameter from
the job definition, while adding the needed parameter.