Correct how we detect nested virtualization layout
Since we have to handle two kinds of layout, multinode and nested
virtualization, we better have to get a strong way to differenciate
them.

Until now, we were trying to be smart, by looking up libvirtd.service
availability, and listing VMs therein. But it wasn't strong enough and
lead to some issues with multinode having libvirtd available.

This patch introduces a new, CI dedicated parameter, allowing to toggle
the right "path" in the log gathering.