Modify libvirt_manager network management
Until now, we were relying on fair luck for the network management and,
especially, the interface attachment: because usually the virtual
network name matches the virtual bridge name, we never faced any issue.

But it may happen this situation isn't the same - for instance the
`default` network has a non-matching bridge name (virbr0), leading to
issues when we want to attach the interface with the `--type bridge`.

Also, with dev-scripts attaching networks using the `--type bridge`, we
must ensure we're really using a bridge name and not a network one.

This patch therefore converge everything to consume:
- bridge type
- bridge name

It will allow to remove install_yamls calls from within the rhol_crc
role and just consume our internal tasks in order to get CRC associated
to the default network.