Add missing uri parameter
While doc states default value is `qemu:///system`, it is not.

Also remove potentially problematic `cached: true`, and add some
debugging tasks, as well as removing the "filter" that may prevent
re-generating the fact if it's already set with some previous run. This
should ensure we're consuming up-to-date list.

References:
- https://docs.ansible.com/ansible/latest/collections/community/libvirt/virt_module.html#parameter-uri
- https://docs.ansible.com/ansible/latest/collections/community/libvirt/virt_net_module.html#parameter-uri