[edpm_prepare]Pass controlplane cr via ansible var
We are working on removing the dependency from install_yamls
reliance. In order to achieve that, we will be setting
NETWORK_ISOLATION to false in Install yamls vars.

In Install yamls, control plane crs are tied with
NETWORK_ISOLATION flag. Since we are going to set it to
false then it will pick different cr leading to breakage in
job.

This pr drops openstack_make_deploy_prep target and allow
users to pass control plane cr via var.

We can use ci_kustomize role to do additional
customization.

Note: We need to kustomize storage class otherwise deployment
will not work.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>