Ensure we have a fancy hostname on CI nodes
Until now, we weren't always sure we were on the correct host, since
their names were meaningless compared to the actual usage.

With this patch, we should end with hostnames matching the actual use,
such as:
- compute-[0-2].ci-rdo.local
- controller.ci-rdo.local

We also lower a bit the wait delay between retries in order to lower the
time to completion for CRC reboot.