Remove Makefile from content provider irrelevant files list
In Install_yamls, we run most of the Zuul jobs based on Makefile changes
and in other operators, Makefile is used to build operator images.

If we put Makefile under content provider ir list, it will skip
content provider job leading to job freeze issue.

Removing it fixes the issue.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>