Allow to create blank VMs
This step is mandatory for baremetal support in the reproducer.

With this patch, setting the "disk_file_name" to "blank" will create a
blank, empty VM. It will also avoid trying to start and configure it (of
course).

In addition, we now generate a MAC address mapping file allowinw to
associate fixed IPs to a MAC and a VM in such a way we can, later,
provision the baremetal nodes properly.