Try to be smarter with zuul jobs
We don't need any deploy job when we affect some of the roles. But we
can't ignore them from the common, base jobs, because some other jobs
are actually consuming them in some way.

This also correct a dependency issue with content-provider not beeing
kicked, while it should for the adoption job.