External projects want to trigger jobs on Makefile
While it makes no sense in the ci-framework project to trigger any
deploy based on Makefile changes, external projects inheriting from our
jobs may want to get this feature, such as install_yamls and
nove_operator.

We may stay as-is, since we don't modify the Makefile that often - and
finding a way out would probably be too much of an issue due to the way
parenting is working.