[edpm_deploy]Install dataplance services
https://github.com/openstack-k8s-operators/install_yamls/pull/650
improves the edpm_deploy_prep target to generate crs with kustomization.

All the oc apply commands are moved to edpm_deploy target. In
CI-framework job, we donot use edpm_deploy target to deploy dataplane.

We use oc apply over dataplane kustomization.
Since dataplane service config[1] are moved to edpm_deploy. So we are
no longer installing these services and it will not start the edpm
compute deployment due to missing services.

In order to fix this, we need to install the services in ci-framework
and it will start the deployment.

Links:
[1]. https://github.com/openstack-k8s-operators/install_yamls/blame/1874191404df509e982ef48699bd5c9f12a5b8f7/Makefile#L680

Signed-off-by: Chandan Kumar <raukadah@gmail.com>