set owner and permissions for libvirt config files
Going forward libvirt will be installed on the host.
https://github.com/openstack-k8s-operators/edpm-ansible/pull/467
As part of that change we now creating the libvirtconfig files
mode "0640". As a result that now causes rsync to fail to copy the
files into the logs. To adress that this change chmods the
libvirt configs to 0644 and recursivly set the ownership.
Closes: #909