Corrections to the reproducer role and doc
The CRC node had some issues with the network, especially the DNS part.
We actually need to reboot it in any cases, so let's just do it.
It will ensure we're using the correct IPs and related data.

We also have to update the reproducer to consume the fixed IP, leading
to a small change in the generated file in order to be expose a
consumable parameter.

Also, the inventory pointed in the doc was wrong.