Enable documentation build for python libs
Since we're getting proper docstring in our python module nested in
module_utils, let's expose them in sphinx.

This requires some changes in the modules directly related to the
custom libs import, to the way we check spelling, and to the doc build
itself, since we have to inject a mid-step, like we did for the
ansible-doc build.

Co-Authored-By: @pablintino