Fix virtual env install
The setup_env script doesn't work as intended with the default values of
using a python virtual environment.

If `pip` is not installed on the host, then it will fail the process.

The issue is that the `case` where the bash script is meant to set the
`PIP` binary to be used is incorrectly written so it doesn't go into the
non virtual env section. This patch fixes it by using a catch all to be
consistent with the `run_molecule` script.

Removing the ' from the values options in the `case` resolves that
problem.

There is another problem caused by `setup_env` using `--user` for `pip
install` but `setup_molecule` isn't using it, so they install in different
locations when not using a virtual env.

Also the gate jobs don't seem to have /root/.local/bin in the PATH, so
we don't use `--user` when running the commands as root.