[ci_kustomize]Allow loading non local files
By default kustomize restrict file loading to directory tree under the
kustomization workspace. However the ci_kustomize action relocates the
kustomization workspace generated by instal_yamls' edpm_deploy_prep
target. That relocation is selective and only moves the manifests but
not other files. So when edpm_deploy_prep generate a congfigMapGenerator
kustomization that references local files in it kustomization workspace
such file is not relocated to the new workspace used by ci_kustomize.
The install_yamls uses absolute file paths in the configMapGenerator so
the relocation does not break the accessibility of the extra files. But
the default behavior of kustomize does not allow loading the file with
the error:

  "error": "loading KV pairs: file sources:
  [25-nova-extra.conf=/home/zuul/ci-framework-data/artifacts/manifests/openstack/dataplane/cr/25-nova-extra.conf]:
  security; file '/home/zuul/ci-framework-data/artifacts/manifests/openstack/dataplane/cr/25-nova-extra.conf'
  is not in or below '/home/zuul/ci-framework-data/artifacts/manifests/openstack/dataplane/cr/cifmw-kustomize-workspace'"

This PR adds `-load-restrictor LoadRestrictionsNone` to the kustomize
call done by ci_kustomize to allow loading the file.