Ensure we gather the various inventories
Since the controller-0 is running on one specific hypervisor, we have to
ensure we can collect and group all of the generated inventories.

This patch ensures we're fetching other hypervisors inventories and copy
them onto the hypervisor running controller-0 before we push everything
onto the controller-0.

That way, we should get all of the needed data for the future steps,
meaning mostly "deploy".

Note that networking may be an issue - but ensuring hypervisors are able
to communicate with each others is not really in the scope of the
ci-framework.