Set right ssh user for crc-0 machine
For convenience we write in ~/.ssh/config the configuration to the
different hosts we can ssh, but we are always writing `zuul` as the
user, which is not correct for `crc`.

This patch makes the user conditional so that the `core` user is used
for the `crc` node and `zuul` for the rest.