Fix idempotency of the push_code task
In the push_code.yml file the reproducer role fetches zuul.items
repositories, unfortunately if we re-run this it will fail, because we
are already in the branch we want to force fetch, which is not allowed.
We need to be in another branch to be able to force-fetch.

This patch checks, for each repository, if the repo directory exists and
if it does it checkouts a different branch so that the fetch doesn't
fail.  For now we only change the branch when job_id is defined.