Change ssh keys to ECDSA
Our current playbook creates and uses ED25519 ssh keys, but those are
not supported by the crypto policy on the compute node, so the
deployment fails.

This patch changes it to be an ECDSA key that will allow the deployment
to complete.  The name of the key is changed from id_ed25519 to id_cifw
so it's not misleading now that we have change the algorithm.

We chose ECDSA instead of other supported algorithms in FIPS to be
consistent with the Nova SSH key.