Be smarter with existing network names
Until now, it was more a "all or none" politic regarding the prefix
added to the network names.

This patch allows to be slightly smarter by allowing to point to actual
network names we don't want to prefix.

This is especially interesting for multi-hypervisor layout, with an OCP
cluster deployed: since ocpbm and ocppr are managed by the 3rd party
product "dev-scripts", this new feature will allow to deploy the same
networks on remote nodes, ensuring consistency in the names all the way
around.