Remove most of the IP management from ansible
All of the IP management will mostly rely on the networking_mapper role
and associated tasks.

In order to avoid duplications, we're now ensuring we're just keeping
the minimal parts to get the network interfaces attached to the VMs;
we're also removing the static_ip support and the support for the
`cifmw_libvirt_manager_networks` parameter.

This patch has been locally tested against:
- 3-nodes.yml reproducer scenario
- validated-architecture-1.yml reproducer scenario