Refactor reproducer role to consume networking_mapper
This patch switches the reproducer role to networking_mapper when it
comes to networking configuration.

It means we have to make a deep refactoring, as well as changing some of
the logic underneath the various tasks.

Mostly, we have to ensure the controller-0 is accessed using the right
user (see previous patch in this same PR), but also ensure controller-0
is ready before we get to the CRC (and OCP) node(s). This is mandatory,
especially for CRC, because we need to get its fixed IP.
This IP is generated via the networking_mapper, so we have to call it
early enough.

Since we'll be consuming some crafted output generated from the
networking_mapper data during the product deployment, we also have to
ensure we have all the needed pieces on the controller-0 - meaning it's
easier to run everything from that node directly, so that the content
gets generated in the right location since the beginning.

There's also a deep cleaning about the delegate_to and related
remote_user, delegate_facts and others since we're consuming a known,
fixed user via the inventory we update dynamically from the
libvirt_manager.

Most of the features should be properly tested from within the molecule
tests running against the reproducer role - but also the ones running
against the libvirt_manager directly.