Ensure hosts are ready before collecting facts
With automation, it may happen hosts are being booted, ansible goes on
other tasks, and when we hit this fact collection, hosts are in a middle
stage where they seem to have some networking issues.

This is especially visible with the OCP cluster - the libvirt_manager
role starts them, wait for them, gets access on them and everything's
good, moves on, but later the OCP cluster is unreachable for a limited
amount of time.

This new task ensure everything's as ready as possible before looping on
the fact collection, avoiding a crash.