Generate ssh keys used for the VMs that tobiko creates
Tobiko creates persistent Openstack workloads. A tobiko container may
need to connect to VMs created by a previous tobiko container.
Due to this, the ssh key files used to connect to those containers have
to be persistent.
With this change, ci-framework generates ssh key files and stores them
on the tobiko artifacts dir, so that they can be reused by different
containers.