Ensure ctlplane interface has an IP for selected nodes
While we were already setting the network on the CRC node in order to
solve some issues, we also have to ensure computes know about it. And,
while at it, ensuring the controller-0 also has that IP ensures we're
able to double-check for the network to function as expected (computes
have a "ping" task against controller-0).