Move creation of clouds.yaml file to edpm_prepare
The clouds.yaml file was created by the tempest role because it is
needed to execute tempest tests.
Later, those tasks from the tempest role were reused by the tobiko role,
because the file is needed for running tobiko as well.
This change moves the creation of clouds.yaml to an earlier step in the
edpm_prepare role and it could be reused by any other stages executed
later (e.g. new test frameworks).